# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple inventory and transactions API",
        "title": "Inventory Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/inventories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of all inventory items from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Get all inventory items",
                "responses": {
                    "200": {
                        "description": "List of all inventories",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new inventory record to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Create a new inventory item",
                "parameters": [
                    {
                        "description": "Inventory Create Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InventoryCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Inventory created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    }
                }
            }
        },
        "/inventories/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve details of a specific inventory item by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Get one inventory item by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Inventory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventory details",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "404": {
                        "description": "Inventory not found",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update details of an existing inventory item by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Update an existing inventory item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Inventory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Inventory Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InventoryUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventory updated successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove an inventory record by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Delete an inventory item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Inventory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventory deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "404": {
                        "description": "Inventory not found",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.GlobalResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.InventoryCreateRequest": {
            "type": "object",
            "required": [
                "item_name",
                "price_per_item",
                "quantity"
            ],
            "properties": {
                "item_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "price_per_item": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "dto.InventoryUpdateRequest": {
            "type": "object",
            "properties": {
                "item_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "price_per_item": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}